Pull color stuff into own file
